quick access - move quick open scorer to a more generic place
