+++ src/vs/workbench/contrib/search/browser/openFileHandler.ts
@@ -13,7 +13,7 @@ import { basename, dirname, toLocalResource } from 'vs/base/common/resources';
 import { URI } from 'vs/base/common/uri';
 import { QuickOpenEntry, QuickOpenModel } from 'vs/base/parts/quickopen/browser/quickOpenModel';
 import { IAutoFocus } from 'vs/base/parts/quickopen/common/quickOpen';
-import { IPreparedQuery, prepareQuery } from 'vs/base/parts/quickopen/common/quickOpenScorer';
+import { IPreparedQuery, prepareQuery } from 'vs/base/common/fuzzyScorer';
 import { IRange } from 'vs/editor/common/core/range';
 import { getIconClasses } from 'vs/editor/common/services/getIconClasses';
 import { IModelService } from 'vs/editor/common/services/modelService';
